% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4health_aggregate.r
\name{data4health_aggregate}
\alias{data4health_aggregate}
\title{Aggregate the dataframe to count frequency}
\usage{
data4health_aggregate(data, space_col, time_col, ...)
}
\arguments{
\item{data}{The dataframe to be filtered. It can be a dataframe from the R
environment, or a filepath (accepted formats are csv, xls, xlsx, dbf,
and dbc).}

\item{space_col}{The name of the column which should be used for the
spatial aggregation.}

\item{time_col}{The name of the column which should be used for the
temporal aggregation}

\item{...}{Any other columns that should be aggregated.}
}
\value{
An aggregated dataframe.
}
\description{
Aggregate the dataframe to count frequency
}
\examples{
data <- data.frame(
  space = c("A", "B", "B", "B"),
  day = c("2023-01-15", "2023-05-20", "2023-01-27", "2023-02-02"),
  month = c("2023-01-01", "2023-05-01", "2023-01-01", "2023-02-02"),
  value1 = c(10, 15, 20, 25),
  value2 = c(5, 8, 12, 17),
  gender = c("Male", "Female", "Female", "Female")
)
data4health_aggregate(data, time_col = "month", space_col = "space", "gender")
}
