% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4health_clean.R
\name{data4health_clean}
\alias{data4health_clean}
\title{Clean the health data}
\usage{
data4health_clean(
  country = NULL,
  disease = NULL,
  data,
  cols_to_remove = NULL,
  cols_to_include = NULL,
  missing_threshold = 100,
  rename = NULL,
  date_to_week = NULL,
  date_to_month = NULL,
  age_threshold = 120,
  output_file,
  save = FALSE
)
}
\arguments{
\item{country}{The country of origin of your health data.}

\item{disease}{The disease that you want to .}

\item{data}{Health data.}

\item{cols_to_remove}{Columns to be removed}

\item{cols_to_include}{Columns to be included}

\item{missing_threshold}{Columns with more  missing values
than this will be excluded}

\item{rename}{A named list of current columns names and how to rename them.}

\item{date_to_week}{Which column should be transformed to week?}

\item{date_to_month}{Which column should be transformed to week?}

\item{age_threshold}{Rows with the age above this number will be removed.}

\item{output_file}{required if save = TRUE. Complete path wher the output of function
should be saved.}

\item{save}{Should the output be saved locally?}
}
\value{
Performs basic cleaning on your files
}
\description{
Clean the health data
}
\examples{
data("peru_dengue_data")
data4health_clean(country = "peru", disease = "dengue",
                       data = peru_dengue_data,
output_file = "/shared/earth/GHR/daniela/harmonize_datapaper/clean_data_output/peru_dengue.csv",
                       cols_to_remove = c("diresa", "ubigeo", "localcod"),
                       rename = c(edade = "age", fecha_nacimiento = "birthdate"),
                       missing_threshold = 20,
                       age_threshold = 120)
}
