% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4health_loadscripts.r
\name{data4health_loadscript}
\alias{data4health_loadscript}
\title{Load the cleaning scripts}
\usage{
data4health_loadscript(country, disease)
}
\arguments{
\item{country}{The country of origin of your health data.}

\item{disease}{The disease that you want to .}
}
\value{
Opens a script in your editor.
}
\description{
Load the cleaning scripts
}
\examples{
data4health_loadscript(country = "colombia", disease = "dengue")
data4health_loadscript(country = "brazil", disease = "malaria")
}
