% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4health_ui.r
\name{data4health_ui}
\alias{data4health_ui}
\title{Load the cleaning scripts}
\usage{
data4health_ui()
}
\value{
Open a webbrowser with the data4health user interface
}
\description{
Load the cleaning scripts
}
\examples{
#data4health_ui()
}
