% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.data4health-package.R
\docType{package}
\name{harmonize.data4health}
\alias{harmonize.data4health}
\title{harmonize.data4health: A Package for Health Data Harmonization}
\description{
This package provides tools for harmonizing health-related data.
}
