% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peru_dengue_data.R
\docType{data}
\name{peru_dengue_data}
\alias{peru_dengue_data}
\title{Example Dataset for peru dengue}
\format{
A data frame with 10 rows and 14 variables:
\describe{
\item{departamento}{Department name}
\item{provincia}{Province name}
\item{distrito}{District name}
\item{localidad}{Locality name}
\item{enfermedad}{Disease name}
\item{ano}{Year of case}
\item{semana}{Week of case}
\item{diagnostic}{Diagnosis result}
\item{diresa}{Health region}
\item{ubigeo}{Geographic code}
\item{localcod}{Local code}
\item{edad}{Age of patient}
\item{tipo_edad}{Type of age (e.g., months, years)}
\item{sexo}{Sex of patient}
}
}
\source{
Simulated data
}
\usage{
peru_dengue_data
}
\description{
A sample dataset for demonstration purposes.
}
\keyword{datasets}
